refactor: moved all piece type tests into sub folder
