+++ src/test/java/domain/piece/KnightTests.java
@@ -1,9 +1,7 @@
-package domain;
+package domain.piece;
 
-import domain.piece.Knight;
-import domain.piece.Pawn;
-import domain.piece.Piece;
-import domain.piece.PieceColor;
+import domain.Board;
+import domain.Location;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
