+++ src/test/java/domain/piece/PieceTests.java
@@ -1,7 +1,6 @@
-package domain;
+package domain.piece;
 
 import org.junit.jupiter.api.Test;
-import domain.piece.*;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
